Add CJK support (for Chinese, Korean and Japanese)
This adds the WenQuanYi Zen Hei font which is licensed under the GPL 2
with font embedding exception, license and author information is
included. The font configuration files of mPDF have been modified in
order to use the font.

The font is automatically selected when a Chinese, Korean or Japanese
locale is set using the "lang" attribute. This attribute is now set
globally based on the configured language. This should work in
combination with the translation plugin, the wrap plugin can be used for
indicating the language of parts of a page.